add: cast

+++ tests/test_clib_math.c
@@ -96,7 +96,7 @@ CHEAT_TEST(
     printf("- Testing %s\n", __func__);
     cheat_assert(clib_math_ctz(0) == 0);
     for (int i = 0; i < 32; i++) {
-        cheat_assert(clib_math_ctz(pow(2, i)) == i);
+        cheat_assert(clib_math_ctz((uint32_t)pow(2, i)) == i);
     })
 
 CHEAT_TEST(
@@ -105,7 +105,7 @@ CHEAT_TEST(
     printf("- Testing %s\n", __func__);
     cheat_assert(clib_math_clz(0) == 31);
     for (int i = 0; i < 32; i++) {
-        cheat_assert(clib_math_clz(pow(2, i)) == (31 - i));
+        cheat_assert(clib_math_clz((uint32_t)pow(2, i)) == (31 - i));
     })
 
 CHEAT_TEST(
@@ -114,7 +114,7 @@ CHEAT_TEST(
     printf("- Testing %s\n", __func__);
     cheat_assert(clib_math_ffs(0) == 0);
     for (int i = 0; i < 32; i++) {
-        cheat_assert(clib_math_ffs(pow(2, i)) == i + 1);
+        cheat_assert(clib_math_ffs((uint32_t)pow(2, i)) == i + 1);
     })
 
 CHEAT_TEST(
